adding swap files to gitignore, splitting into header and source files
